Fix #2320, reorganize EVS header files
Organize EVS header file contents according to the convention
