+++ modules/evs/config/default_cfe_evs_extern_typedefs.h
@@ -25,16 +25,6 @@
 #ifndef CFE_EVS_EXTERN_TYPEDEFS_H
 #define CFE_EVS_EXTERN_TYPEDEFS_H
 
-/* This header may be generated from an EDS file,
- * tools are available and the feature is enabled */
-#ifdef CFE_EDS_ENABLED_BUILD
-
-/* Use the EDS generated version of these types */
-#include "cfe_evs_eds_typedefs.h"
-
-#else
-/* Use the local definitions of these types */
-
 #include "common_types.h"
 
 /**
@@ -166,7 +156,5 @@ enum CFE_EVS_EventOutput
  * @sa enum CFE_EVS_EventOutput
  */
 typedef uint8 CFE_EVS_EventOutput_Enum_t;
-
-#endif /* CFE_EDS_ENABLED_BUILD */
 
 #endif /* CFE_EVS_EXTERN_TYPEDEFS_H */
