+++ modules/evs/config/default_cfe_evs_msgdefs.h
@@ -18,31 +18,26 @@
 
 /**
  * @file
+ *   Specification for the CFE Event Services (CFE_EVS) command and telemetry
+ *   message constant definitions.
  *
- * Encapsulates all EVS module internal header files, as well
- * as the public API from all other CFE core modules, OSAL, and PSP.
- *
- * This simplifies the set of include files that need to be put at the
- * start of every source file.
+ *  For CFE_EVS this is only the function/command code definitions
  */
+#ifndef CFE_EVS_MSGDEFS_H
+#define CFE_EVS_MSGDEFS_H
 
-#ifndef CFE_EVS_MODULE_ALL_H
-#define CFE_EVS_MODULE_ALL_H
+#include "cfe_evs_fcncodes.h"
 
-/********************* Include Files  ************************/
+/* Event Type bit masks */
+#define CFE_EVS_DEBUG_BIT       0x0001
+#define CFE_EVS_INFORMATION_BIT 0x0002
+#define CFE_EVS_ERROR_BIT       0x0004
+#define CFE_EVS_CRITICAL_BIT    0x0008
 
-#include "cfe.h" /* All CFE+OSAL public API definitions */
-#include "cfe_platform_cfg.h"
+/* Output Port bit masks */
+#define CFE_EVS_PORT1_BIT 0x0001
+#define CFE_EVS_PORT2_BIT 0x0002
+#define CFE_EVS_PORT3_BIT 0x0004
+#define CFE_EVS_PORT4_BIT 0x0008
 
-#include "cfe_msgids.h"
-#include "cfe_perfids.h"
-
-#include "cfe_evs_core_internal.h"
-
-#include "cfe_evs_eventids.h" /* EVS event IDs */
-#include "cfe_evs_task.h"     /* EVS internal definitions */
-#include "cfe_evs_log.h"      /* EVS log file definitions */
-#include "cfe_evs_utils.h"    /* EVS utility function definitions */
-#include "cfe_evs_dispatch.h"
-
-#endif /* CFE_EVS_MODULE_ALL_H */
+#endif
